services/simulator/requirements.txt
pyais==2.9.1
websockets==13.1
pyyaml==6.0.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
